pyproject.toml
[project]
name = "dev-bot"
version = "0.1.0"
description = "Autonomous developer bot using Claude Agent SDK"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk",
    "filelock",
    "httpx",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]

[project.scripts]
dev-bot = "bot.run:main"

[tool.uv.workspace]
members = ["memory-server"]
